Updated GUI Header
Updated the GUI header to accept a vector of functions - the functions are run by the server depending on what the first 3 bytes of the received message (from the JS script) is. The string that the function returns is what is sent back to the JS script - this is the core of the GUI interface, developers should create their own custom functions to add to this input vector.